x/medichain/keeper/msg_server_admin_test.go
package keeper

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"

	"github.com/sota/medichain/x/medichain/types"
)

func TestAdminMsgServerCreate(t *testing.T) {
	srv, ctx := setupMsgServer(t)
	creator := "A"
	for i := 0; i < 5; i++ {
		resp, err := srv.CreateAdmin(ctx, &types.MsgCreateAdmin{Creator: creator})
		require.NoError(t, err)
		assert.Equal(t, i, int(resp.Id))
	}
}

func TestAdminMsgServerUpdate(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgUpdateAdmin
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgUpdateAdmin{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateAdmin{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateAdmin{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		tc := tc
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)
			_, err := srv.CreateAdmin(ctx, &types.MsgCreateAdmin{Creator: creator})
			require.NoError(t, err)

			_, err = srv.UpdateAdmin(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}

func TestAdminMsgServerDelete(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgDeleteAdmin
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgDeleteAdmin{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgDeleteAdmin{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "KeyNotFound",
			request: &types.MsgDeleteAdmin{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		tc := tc
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)

			_, err := srv.CreateAdmin(ctx, &types.MsgCreateAdmin{Creator: creator})
			require.NoError(t, err)
			_, err = srv.DeleteAdmin(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}
